fix: additional cleanup
