+++ .pre-commit-config.yaml
@@ -51,10 +51,18 @@ repos:
             oracledb,
             psycopg,
             pytest,
+            pytest-asyncio,
+            pytest-lazy-fixture,
+            time-machine,
+            httpx,
+            "sanic[ext]",
+            socketify,
+            fastapi,
             pytest-mock,
             pytest-timeout,
             sqlalchemy>=2.0.12,
             msgspec,
             click,
-            litestar,
+            rich,
+            "litestar[cli]",
           ]
